move or_json encoders to their own file
